requirements.txt
obs-websocket-py>=5.0.0
tkinter>=8.6
requests>=2.25.1
python-dotenv>=0.19.0
Flask==2.0.1
websockets==12.0
